+++ crates/prelude/CHANGELOG.md
@@ -1,13 +1,20 @@
 # Changelog
 
-## 0.7.1-git
+## 0.8.0-git
+
+### Major
+
+- Make `syscall()` unsafe to call
 
 ### Minor
 
+- Support the new behavior of the scheduler (not allocating if the size is zero)
+- Add `rng::bytes{,_array}()` as safer alternatives to `rng::fill_bytes()`
 - Use Rust edition 2024
 
 ### Patch
 
+- Fix clippy lints
 - Update dependencies
 
 ## 0.7.0
@@ -185,4 +192,4 @@
 
 ## 0.1.0
 
-<!-- Increment to skip CHANGELOG.md test: 1 -->
+<!-- Increment to skip CHANGELOG.md test: 2 -->